Development (#45)
* Update node to 16.16 & run `npm update`

Due to an update in the default repositories in the underlying alpine image, python no longer refers to the correct version for our build. Changing it to python3 will allow the build to complete succesfully. See https://stackoverflow.com/questions/62169568/docker-alpine-linux-python-missing

Ideally we'd update to version 18 straight away, but that seems a bit less straightforward. Updating to version 16 (which has security support until september 11th 2023) buys us time to come up with a good strategy to upgrade further.

This also updates the `yarn.lock` file.

* npm audit fix

* Add github templates

* Fix the breaking changes of bookschelf

* Add `MYSQL_CA_CERT` for MySQL SSL connection

* Update changelog

* Changed docker file to create image on commit

* fix: make use of docker image cache, fix knex error, optimize image

* Remove bookshelf/knex, add Sequelize

* typo

* Update version

* Remove trailing slash

* npm update

* Update sequelize implementation and add seed data

---------

Co-authored-by: Rudi van Hierden <rudi@draad.nl>
Co-authored-by: Niels Vegter <niels@denes.nl>
Co-authored-by: Daan Rosbergen <d.rosbergen@gmail.com>
Co-authored-by: Daan Rosbergen <Badmuts@users.noreply.github.com>